Moved all panic stuff into a panic module
